artur-tg-bot-1-master/requirements.txt
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
et-xmlfile==1.1.0
Flask==2.3.2
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.3
MarkupSafe==2.1.3
openpyxl==3.1.2
Pillow==10.0.0
psycopg2==2.9.7
pyTelegramBotAPI==4.12.0
python-docx==0.8.11
requests==2.31.0
SQLAlchemy==2.0.19
telebot==0.0.5
typing_extensions==4.7.1
urllib3==2.0.4
Werkzeug==2.3.6
